Enable editing of ACL constants
Add an editing_key signal and branch create vs update based on it.
Pre-fill the form when editing and show "Update constant" plus a "Cancel
edit" button. Wire AclRow to accept an on_edit handler and add an Edit
button to populate the form. Add #[allow(unused_imports)] in the ACL
store mod.